Updated to Rails 3.2.21
Note: to run rspec or cucumber, the condor3_test database must be a copy
of the condor3_template database.  The way to create this is to
uncomment the line:

  #  template: condor3_template

In the test stanza of config/database.yml.  Then do:

RAILS_ENV=test rake db:drop db:create

Then comment out the template line (put it back like it was).  At this
point, cucumber and rspec should pass.