Got rspec working etc.
Wrote update in tools/database_state/README to provide instructions on
how to create the template database and pulled out git log entry on
how to use it.

Got rspec working with no errors and no deprecations.

Can now do rake assets:precompile which appears to be needed for one
rspec test.

cucumber appears to work except for the ones that need capybara-webkit.