Check point on my quest to deploy to new server
This branch is for the version that is deployed on tcp149.  I bailed
out of using thin for development and Passenger for production and
switched to using Puma.  The current state is it is working on tcp149
but the javascript is not kicking off and formatting the
upd_apar_defs.  This check point I have development mode working using
Puma.